Merge commit '02bc8a885a5553b258a35cd8d10c8393441c9afe'
Shots are now target tracking and hits a creature object in the middle instead of down to the left.

An option activity has been created where the player browses between three different maps. The choice is sendid to the game loop via an intent so that the right configurations from data file can be loaded.
Also option for level choice, although not implemented yet

Have removed the title bar from all activities.

Made an intro sequence for the application, with time duration. It also has fade in and fade out animations, though the animations are only supported from version 2.0 so they are comments for now.