modified:   jni/render. modified:   libs/armeabi/librender.so modified:   src/com/crackedcarrot/GameLoop.java modified:   src/com/crackedcarrot/NativeRender.java
This takes care of the uber evil render bugs.. nothing that complex really.
We screwed ourselves pretty big when i redid the render and lots of the old
code took advantage of the very very abuse friendly old code.

NOTE TO EVERYONE, SET THE SIZE OF SPRITES BEFORE YOU CALL finalizeSprites() !!

and yes i will implement the capability to change size on the fly during a level.
just don't forget this for now.

Oh and fun todo for anyone who wants to beat themselves.. change all member variables
that are not accessed frequently to private and implement getters.

The way it is now with tons and tons of public variables that are used extremely infrequently
or only during startup SUCKS!